+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.4.0 - [2025-10-]
+
+### `Added`
+
+### `Changed`
+
+### `Fixed`
+
+### `Deprecated`
+- [#13](https://github.com/agusinac/OmicFlow/issues/13) Removed `viridis` dependency, `purrr:map` is replaced by `lapply`
+
+## v1.3.2 - [2025-09-20]
+
+### `Deprecated`
+- `testthat` removed autoflow test
+
 ## v1.3.1 - [2025-09-03]
 
 ### `Added`